Create apple-maps example
