+++ examples/get-started/pure-js/apple-maps/README.md
@@ -1,0 +1,21 @@
+## Example: Use deck.gl with Apple Maps
+
+Uses [Vite](https://vitejs.dev/) to bundle and serve files.
+
+## Usage
+
+To run this example, you need a [Apple Maps token](https://developer.apple.com/documentation/mapkitjs/creating_a_maps_token). 
+
+You need to then set the `data-token` property in `index.html` to this token.
+
+To install dependencies:
+
+```bash
+npm install
+# or
+yarn
+```
+
+Commands:
+* `npm start` is the development target, to serve the app and hot reload.
+* `npm run build` is the production target, to create the final bundle and write to disk.
