+++ examples/get-started/pure-js/apple-maps/index.html
@@ -1,0 +1,41 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset='UTF-8' />
+    <title>deck.gl w/ Apple Maps example</title>
+    <style>
+      body {
+        margin: 0;
+        padding: 0;
+        font-family: UberMove, Helvetica, Arial, sans-serif;
+      }
+      #container {
+        position: fixed;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+      }
+      #container > * {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+      }
+    </style>
+  </head>
+  <script src="https://cdn.apple-mapkit.com/mk/5.x.x/mapkit.core.js"
+          crossorigin async
+          data-callback="initMapKit"
+          data-libraries="full-map"
+          data-token="XXX"
+  ></script>
+  <body>
+    <div id="container">
+      <div id="map"></div>
+      <canvas id="deck-canvas"></canvas>
+    </div>
+    <script type="module" src='app.js'></script>
+  </body>
+</html>
