+++ examples/get-started/pure-js/google-maps/index.html
@@ -2,21 +2,9 @@
 <html>
   <head>
     <meta charset='UTF-8' />
-    <title>deck.gl w/ Apple Maps example</title>
+    <title>deck.gl w/ Google Maps example</title>
     <style>
-      body {
-        margin: 0;
-        padding: 0;
-        font-family: UberMove, Helvetica, Arial, sans-serif;
-      }
-      #container {
-        position: fixed;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-      }
-      #container > * {
+      #map {
         position: absolute;
         top: 0;
         left: 0;
@@ -25,17 +13,8 @@
       }
     </style>
   </head>
-  <script src="https://cdn.apple-mapkit.com/mk/5.x.x/mapkit.core.js"
-          crossorigin async
-          data-callback="initMapKit"
-          data-libraries="full-map"
-          data-token="XXX"
-  ></script>
   <body>
-    <div id="container">
-      <div id="map"></div>
-      <canvas id="deck-canvas"></canvas>
-    </div>
+    <div id="map"></div>
     <script type="module" src='app.js'></script>
   </body>
 </html>
